+++ lib/lite/ruby/safe/range.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Range
+
+  def overlaps?(other)
+    cover?(other.first) || other.cover?(first)
+  end
+
+end