+++ spec/lite/ruby/integer_spec.rb
@@ -110,7 +110,7 @@ RSpec.describe Integer do
 
   describe '#to_time' do
     it 'to be "1969-12-31 19:00:03 -0500"' do
-      expect(3.to_time.to_s).to be_a(String)
+      expect(3.to_time).to be_a(Time)
     end
   end